onended callback registration
