+++ audio/src/decoder.rs
@@ -1,5 +1,6 @@
 use std::boxed::FnBox;
 use std::sync::Mutex;
+
 pub struct AudioDecoderCallbacks {
     pub eos: Mutex<Option<Box<FnBox() + Send + 'static>>>,
     pub error: Mutex<Option<Box<FnBox() + Send + 'static>>>,
